lma/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.62
botocore==1.17.62
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.9
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
idna==2.10
isort==5.5.2
jmespath==0.10.0
lazy-object-proxy==0.0.0
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.51
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.51.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
